lib: libcontainer references are linux only
